Tests/LibWeb/Text/expected/DOM/window-open-and-reopen-custom-target.txt

1: Target matches source: true
2: Target matches source: true
